+++ aleph-client/src/lib.rs
@@ -6,12 +6,11 @@ pub use balances::total_issuance;
 use codec::{Decode, Encode};
 pub use debug::print_storages;
 pub use elections::{
-    change_ban_config, get_ban_config, get_ban_reason_for_validator, get_committee_seats,
+    change_ban_config, get_ban_config, get_ban_info_for_validator, get_committee_seats,
     get_current_era_non_reserved_validators, get_current_era_reserved_validators,
     get_current_era_validators, get_era_validators, get_next_era_committee_seats,
     get_next_era_non_reserved_validators, get_next_era_reserved_validators,
     get_next_era_validators, get_underperformed_validator_session_count, get_validator_block_count,
-    kick_out_from_committee, set_kick_out_config,
 };
 pub use fee::get_next_fee_multiplier;
 pub use finalization::set_emergency_finalizer as finalization_set_emergency_finalizer;
