+++ e2e-tests/src/cases.rs
@@ -2,7 +2,8 @@ use crate::{
     config::Config,
     test::{
         authorities_are_staking as test_authorities_are_staking,
-        ban_automatic as test_ban_automatic, batch_transactions as test_batch_transactions,
+        ban_automatic as test_ban_automatic, ban_threshold as test_ban_threshold,
+        batch_transactions as test_batch_transactions,
         change_stake_and_force_new_era as test_change_stake_and_force_new_era,
         change_validators as test_change_validators,
         channeling_fee_and_tip as test_channeling_fee_and_tip,
@@ -11,7 +12,6 @@ use crate::{
         fee_calculation as test_fee_calculation, finalization as test_finalization,
         force_new_era as test_force_new_era, points_basic as test_points_basic,
         points_stake_change as test_points_stake_change,
-        points_basic as test_points_basic, points_stake_change as test_points_stake_change,
         schedule_version_change as test_schedule_version_change,
         staking_era_payouts as test_staking_era_payouts,
         staking_new_validator as test_staking_new_validator, token_transfer as test_token_transfer,
@@ -28,7 +28,7 @@ pub type PossibleTestCases = Vec<(&'static str, TestCase)>;
 /// This comes up in local tests.
 pub fn possible_test_cases() -> PossibleTestCases {
     vec![
-        ("ban_threshold", test_kick_out_threshold as TestCase),
+        ("ban_threshold", test_ban_threshold as TestCase),
         ("finalization", test_finalization as TestCase),
         ("version_upgrade", test_schedule_version_change),
         ("rewards_disable_node", test_disable_node as TestCase),
@@ -60,11 +60,10 @@ pub fn possible_test_cases() -> PossibleTestCases {
             "authorities_are_staking",
             test_authorities_are_staking as TestCase,
         ),
+        ("ban_automatic", test_ban_automatic as TestCase),
         (
             "clearing_session_count",
             test_clearing_session_count as TestCase,
         ),
-        ("ban_automatic", test_ban_automatic as TestCase),
-        ("ban_manual", test_kick_out_manual as TestCase),
     ]
 }
