+++ e2e-tests/src/test/mod.rs
@@ -1,4 +1,4 @@
-pub use ban::{ban_automatic, ban_manual, clearing_session_count, ban_threshold};
+pub use ban::{ban_automatic, ban_threshold, clearing_session_count};
 pub use electing_validators::authorities_are_staking;
 pub use era_payout::era_payouts_calculated_correctly;
 pub use era_validators::era_validators;
